+++ app.go
@@ -36,7 +36,10 @@ import (
 	"go.uber.org/multierr"
 )
 
-const defaultTimeout = 15 * time.Second
+// DefaultTimeout is the default timeout for starting or stopping an
+// application. It can be configured with the StartTimeout and StopTimeout
+// options.
+const DefaultTimeout = 15 * time.Second
 
 // An Option configures an App using the functional options paradigm
 // popularized by Rob Pike. If you're unfamiliar with this style, see
@@ -179,6 +182,20 @@ func (og optionGroup) String() string {
 	return fmt.Sprintf("fx.Options(%s)", strings.Join(items, ", "))
 }
 
+// StartTimeout changes the application's start timeout.
+func StartTimeout(v time.Duration) Option {
+	return optionFunc(func(app *App) {
+		app.startTimeout = v
+	})
+}
+
+// StopTimeout changes the application's stop timeout.
+func StopTimeout(v time.Duration) Option {
+	return optionFunc(func(app *App) {
+		app.stopTimeout = v
+	})
+}
+
 // Printer is the interface required by Fx's logging backend. It's implemented
 // by most loggers, including the one bundled with the standard library.
 type Printer interface {
@@ -238,12 +255,14 @@ func (l nopLogger) Printf(string, ...interface{}) {
 // execute one at a time, in reverse order, and must all complete within a
 // configurable deadline (again, 15 seconds by default).
 type App struct {
-	err       error
-	container *dig.Container
-	lifecycle *lifecycleWrapper
-	provides  []interface{}
-	invokes   []interface{}
-	logger    *fxlog.Logger
+	err          error
+	container    *dig.Container
+	lifecycle    *lifecycleWrapper
+	provides     []interface{}
+	invokes      []interface{}
+	logger       *fxlog.Logger
+	startTimeout time.Duration
+	stopTimeout  time.Duration
 }
 
 // New creates and initializes an App, immediately executing any functions
@@ -254,9 +273,11 @@ func New(opts ...Option) *App {
 	lc := &lifecycleWrapper{lifecycle.New(logger)}
 
 	app := &App{
-		container: dig.New(),
-		lifecycle: lc,
-		logger:    logger,
+		container:    dig.New(),
+		lifecycle:    lc,
+		logger:       logger,
+		startTimeout: DefaultTimeout,
+		stopTimeout:  DefaultTimeout,
 	}
 
 	for _, opt := range opts {
@@ -281,14 +302,14 @@ func New(opts ...Option) *App {
 }
 
 // Run starts the application, blocks on the signals channel, and then
-// gracefully shuts the application down. It uses the default value of 15
-// seconds for the start and stop timeouts. It's designed to make typical
-// applications simple to run.
+// gracefully shuts the application down. It uses DefaultTimeout to set a
+// deadline for application startup and shutdown, unless the user has
+// configured different timeouts with the StartTimeout or StopTimeout options.
+// It's designed to make typical applications simple to run.
 //
 // However, all of Run's functionality is implemented in terms of the exported
 // Start, Done, and Stop methods. Applications with more specialized needs
-// (including custom start and stop timeouts) can use those methods directly
-// instead of relying on Run.
+// can use those methods directly instead of relying on Run.
 func (app *App) Run() {
 	app.run(app.Done())
 }
@@ -346,6 +367,20 @@ func (app *App) Done() <-chan os.Signal {
 	return c
 }
 
+// StartTimeout returns the configured startup timeout. Apps default to using
+// DefaultTimeout, but users can configure this behavior using the
+// StartTimeout option.
+func (app *App) StartTimeout() time.Duration {
+	return app.startTimeout
+}
+
+// StopTimeout returns the configured shutdown timeout. Apps default to using
+// DefaultTimeout, but users can configure this behavior using the StopTimeout
+// option.
+func (app *App) StopTimeout() time.Duration {
+	return app.stopTimeout
+}
+
 func (app *App) provide(constructor interface{}) {
 	if app.err != nil {
 		return
@@ -388,7 +423,7 @@ func (app *App) executeInvokes() error {
 }
 
 func (app *App) run(done <-chan os.Signal) {
-	startCtx, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+	startCtx, cancel := context.WithTimeout(context.Background(), app.StartTimeout())
 	defer cancel()
 
 	if err := app.Start(startCtx); err != nil {
@@ -397,7 +432,7 @@ func (app *App) run(done <-chan os.Signal) {
 
 	app.logger.PrintSignal(<-done)
 
-	stopCtx, cancel := context.WithTimeout(context.Background(), defaultTimeout)
+	stopCtx, cancel := context.WithTimeout(context.Background(), app.StopTimeout())
 	defer cancel()
 
 	if err := app.Stop(stopCtx); err != nil {