+++ fxtest/fxtest.go
@@ -64,7 +64,10 @@ func New(tb TB, opts ...fx.Option) *App {
 
 // RequireStart calls Start, failing the test if an error is encountered.
 func (app *App) RequireStart() *App {
-	if err := app.Start(context.Background()); err != nil {
+	startCtx, cancel := context.WithTimeout(context.Background(), app.StartTimeout())
+	defer cancel()
+
+	if err := app.Start(startCtx); err != nil {
 		app.tb.Errorf("application didn't start cleanly: %v", err)
 		app.tb.FailNow()
 	}
@@ -73,7 +76,10 @@ func (app *App) RequireStart() *App {
 
 // RequireStop calls Stop, failing the test if an error is encountered.
 func (app *App) RequireStop() {
-	if err := app.Stop(context.Background()); err != nil {
+	stopCtx, cancel := context.WithTimeout(context.Background(), app.StopTimeout())
+	defer cancel()
+
+	if err := app.Stop(stopCtx); err != nil {
 		app.tb.Errorf("application didn't stop cleanly: %v", err)
 		app.tb.FailNow()
 	}